Restructure Album Page-related components and styles
