+++ src/components/album/content/AlbumReviewList.js
@@ -1,0 +1,11 @@
+/* @flow */
+
+import React from 'react';
+
+const AlbumReviewList = () => (
+    <div>
+        <h1>Review list.</h1>
+    </div>
+);
+
+export default AlbumReviewList;
