+++ src/components/album/summary/AlbumMenu.css
@@ -1,0 +1,25 @@
+.album-menu {
+    position: relative;
+}
+
+.album-menu__button {
+    background-color: transparent;
+    border-radius: 0;
+    margin: 0;
+    padding: .5rem;
+    display: block;
+    width: 100%;
+    text-transform: uppercase;
+}
+
+.album-menu__button:hover {
+    background-color: rgba(28, 37, 155, 0.8);
+}
+
+.album-menu__button--play {
+    background-color: rgb(30, 215, 96, 0.5);
+}
+
+.album-menu__button--play:hover {
+    background-color: rgb(30, 215, 96, 0.8);
+}
