+++ src/components/album/summary/ArtistAlbumsGrid.css
@@ -1,0 +1,70 @@
+.artist-albums-grid {
+    text-transform: uppercase;
+}
+
+.artist-albums-grid__title {
+    padding-top: 1rem;
+    padding-bottom: 1rem;
+    margin-bottom: 0;
+    text-align: center;
+}
+
+.artist-albums-grid__content {
+    margin-left: 0;
+    margin-right: .5rem;
+    margin-bottom: -.5rem;
+}
+
+.artist-album-grid__item {
+    padding: 0 0 .5rem .5rem;
+}
+
+.more-albums-button {
+    background-size: 100%;
+    height: 100%;
+    width: 100%;
+    position: relative;
+    transition: background-image .15s ease-in-out;
+}
+
+.more-albums-button__content {
+    background-color: rgba(0, 0, 0, 0.7);
+    height: 100%;
+    width: 100%;
+    position: absolute;
+    transition: color .15s ease-in-out, background-color .15s ease-in-out;
+}
+
+.more-albums-button__content:hover {
+    background-color: rgba(0, 0, 0, 0.9);
+    color: white;
+    cursor: pointer;
+}
+
+.artist-album, .more-albums-button__link {
+    width: inherit;
+}
+
+.more-albums-button__link {
+    white-space: normal;
+    width: 100%;
+    padding: 1rem;
+    margin: 0;
+    position: absolute;
+    top: 50%;
+    left: 50%;
+    transform: translate(-50%, -50%);
+    font-size: 3rem;
+}
+
+.artist-album:hover {
+    cursor: pointer;
+}
+
+.artist-album:hover, .artist-album--current {
+    border: .25rem solid whitesmoke;
+}
+
+.artist-album--current:hover {
+    cursor: default;
+}
